% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_correct.R
\name{BiasCorrect}
\alias{BiasCorrect}
\title{Corrects for bias in parameter estimates due to model uncertainty.}
\usage{
BiasCorrect(n, true, param, models, N = 1000, ic = "AIC")
}
\arguments{
\item{n:}{Sample size.}

\item{true:}{A string giving the name of the model to generate data from.}

\item{models:}{A string vector of model names for selecting between.}

\item{param:}{A vector of parameters specifying a fitted true model.}

\item{N:}{Number of bootstrap samples.}

\item{ic:}{String giving which information criterion to use (e.g. 'AICc', 'BIC')}
}
\value{
: The corrected model parameters.
}
\description{
Corrects for bias in parameter estimates due to model uncertainty.
}
\examples{
data = GenerateData(25, "norm", c(mu = 0.3, sd = 1.2))
models = c("norm0", "norm1", "norm")
ecic = ECIC(data, models, alpha = 0.05, N = 100, ic = "AIC", correct = T)

}
