% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecic.R
\name{ECIC}
\alias{ECIC}
\title{Perform the Error Control for Information Criteria (ECIC) Procedure.}
\usage{
ECIC(models, data, alpha = c(0.01, 0.05, 0.1), N = 1000, ic = "AIC")
}
\arguments{
\item{data:}{A vector/matrix data samplecompatible with the specified models.}

\item{model:}{A string vector specifying the models to use.}
}
\value{
An ecic object with the following attributes:
   \code{decisions:  }A dataframe containing the results of the ECIC procedure, along
   with the decisions given by the best scoring model and the Burnham &
   Anderson (BA) rule (accept best model if delta < -10).

   \code{observed:  } an ecicObserved object containing the observed IC delta,
   the information criterion values for each model given the data, fitted
   parameters for each model, and the observed data itself.

   \code{bias.correction:  }An ecicBias object containing the corrected model
   parameters and the data used for the bias correction procedure.

   \code{error.control:  } A list of ecicControl objects corresponding to each
   alternative model. Each ecicControl object contains information related
   to the error control bootstrap procedure: a vector of model bootstrap
   frequencies, a vector of bootstrap IC deltas, estimated parameters
   corresponding to each bootstrap sample, a matrix of model scores, and the
   bootstrap data itself.
}
\description{
Perform the Error Control for Information Criteria (ECIC) Procedure.
}
\examples{
my.data = GenerateData(25, "norm", c(mu = 0.3, sd = 1.2))
my.models = c("norm0", "norm1" "norm")
ECIC(my.data, my.models, alpha = 0.05, N = 1000, ic = "AIC")

}
