% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecic_control.R
\name{ecicControl}
\alias{ecicControl}
\title{Bootstrap distributions of Information Criterion Differences}
\usage{
ecicControl(n, true, parameters, best, models, N = 1000, ic = "AIC")
}
\arguments{
\item{n:}{Sample size.}

\item{true:}{A string giving the name of the model to generate data from.}

\item{param:}{A vector of parameters specifying a fitted true model.}

\item{best:}{A string giving the name of the model to be given as best.}

\item{models:}{A string vector of model names for selecting between.}

\item{N:}{Number of bootstrap samples.}

\item{ic:}{String giving which information criterion to use (e.g. 'AICc', 'BIC')}
}
\value{
An ecicControl object containing the following:

  differences:

  frequencies: a modelFrequencies object containing bootstrap frequencies,
  fitted parameters for the bootstrap samples, scores for the bootstrap samples
  and the samples themselves.
}
\description{
Bootstrap distributions of Information Criterion Differences
}
\examples{
my.models = c("norm0" "norm1", "norm")
ecicControl(n = 25, true = "norm", param = c(0.3, 1.2),
            my.models, N = 1000, ic = 'AIC')

}
