% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters.R
\name{EstimateParameters}
\alias{EstimateParameters}
\title{Computes the fitted parameters for a data sample and a model.}
\usage{
EstimateParameters(model, data)
}
\arguments{
\item{data:}{compatible with the specified model.}

\item{model:}{A string specifying the model to compute the log-likehood for.}
}
\value{
A vector/matrix of the fitted parameters.
}
\description{
Computes the fitted parameters for a data sample and a model.
}
\examples{
data = GenerateData(25, "norm", c(mu = 0.3, sd = 1.2))
models = c("norm0", "norm1", "norm")
EstimateParameters(data)
}
