% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters.R
\name{EstimateParametersMulti}
\alias{EstimateParametersMulti}
\title{Computes the fitted parameters for multiple data samples and a model.}
\usage{
EstimateParametersMulti(data, model, flat = F)
}
\arguments{
\item{data:}{A matrix of data samples compatible with the specified model.}

\item{model:}{A string specifying the model to compute the log-likehood for.}
}
\value{
A vector/matrix of the fitted parameters.
}
\description{
Computes the fitted parameters for multiple data samples and a model.
}
\examples{
data = GenerateDataMulti(25, "norm", c(mu = 0.3, sd = 1.2), 10)
models = c("norm0", "norm1", "norm")
EstimateParameters(data)

}
