% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{GenerateData}
\alias{GenerateData}
\title{Generate a sample from a fitted model}
\usage{
GenerateData(n, model, parameters = list())
}
\value{
: A list containing the following:.
}
\description{
Generate a sample from a fitted model
}
\examples{
GenerateData(25, "norm", c(mu = 0.3, sd = 1.2))

}
