% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{GenerateDataBest}
\alias{GenerateDataBest}
\title{Generate multiple samples from a fitted model with specified best performing
model}
\usage{
GenerateDataBest(n, true, parameters, best, models, N, ic = "AIC", ...)
}
\arguments{
\item{n:}{Sample size.}

\item{true:}{A string giving the name of the model to generate data from.}

\item{param:}{A vector of parameters specifying a fitted true model.}

\item{best:}{A string giving the name of the model to be given as best.}

\item{models:}{A string vector of model names for selecting between.}

\item{N:}{Number of bootstrap samples.}

\item{ic:}{String giving which information criterion to use (e.g. 'AICc', 'BIC')}
}
\value{
: A list containing the following:.
}
\description{
Generate multiple samples from a fitted model with specified best performing
model
}
\examples{
GenerateData(25, "norm", c(mu = 0.3, sd = 1.2))
}
