% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{GenerateDataMulti}
\alias{GenerateDataMulti}
\title{Generate multiple samples from a fitted model}
\usage{
GenerateDataMulti(n, model, parameters, N)
}
\value{
: A list containing the following:.
}
\description{
Generate multiple samples from a fitted model
}
\examples{
GenerateData(25, "norm", c(mu = 0.3, sd = 1.2))
}
