% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_scores.R
\name{IC}
\alias{IC}
\title{Computes a given information criterion for data and a specified model.}
\usage{
IC(data, model, ic = "AIC", compress = F)
}
\arguments{
\item{data:}{A data sample compatible witih the model.}

\item{model:}{A string containing the model name.}

\item{ic:}{String giving  information criterion to use (e.g. 'AICc', 'BIC')}
}
\value{
: A list containing the following:.
}
\description{
Computes a given information criterion for data and a specified model.
}
\examples{


}
