% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_scores.R
\name{ICMulti}
\alias{ICMulti}
\title{Computes an information criterion for many data samples and specified model.}
\usage{
ICMulti(data, model, ic = "AIC")
}
\arguments{
\item{n:}{Sample size.}

\item{model:}{A string containing the model name.}

\item{ic:}{String giving information criterion to use (e.g. 'AICc', 'BIC')}
}
\value{
: A list containing the following:.
}
\description{
Computes an information criterion for many data samples and specified model.
}
\examples{


}
