% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_scores.R
\name{ICMultiMulti}
\alias{ICMultiMulti}
\title{Computes an information criterion for many data samples and list of models.}
\usage{
ICMultiMulti(models, data, ic = "AIC", compress = F)
}
\arguments{
\item{n:}{Sample size.}

\item{model:}{A vector of strings containing the model names.}

\item{ic:}{String giving information criterion to use (e.g. 'AICc', 'BIC')}
}
\value{
: A list containing the following:.
}
\description{
Computes an information criterion for many data samples and list of models.
}
\examples{


}
