% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{logLik.ecicModel}
\alias{logLik.ecicModel}
\title{Computes log-likelihood for data and a specified model.}
\usage{
\method{logLik}{ecicModel}(model, data, compress = F)
}
\arguments{
\item{model:}{An ecicModel object.}

\item{data:}{A data sample compatible with the model.}
}
\value{
: A list containing the following:.
}
\description{
Computes log-likelihood for data and a specified model.
}
\examples{

}
