% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{loglik}
\alias{loglik}
\title{Computes log-likelihood for data and a specified model.}
\usage{
loglik(data, model, compress = F)
}
\arguments{
\item{data:}{A data sample compatible witih the model.}

\item{model:}{A string containing the model name.}

\item{ic:}{String giving  information criterion to use (e.g. 'AICc', 'BIC')}
}
\value{
: A list containing the following:.
}
\description{
Computes log-likelihood for data and a specified model.
}
\examples{

}
