% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{logLikMulti}
\alias{logLikMulti}
\title{Computes log-likelihood for many data samples and specified model.}
\usage{
logLikMulti(model, data, compress = F)
}
\arguments{
\item{data:}{A matrix of data samples compatible with the model.}

\item{model:}{A string containing the model name.}

\item{ic:}{String giving information criterion to use (e.g. 'AICc', 'BIC')}
}
\value{
: A list containing the following:.
}
\description{
Computes log-likelihood for many data samples and specified model.
}
\examples{
data = GenerateDataMulti(25, "norm", c(0.3, 1.2), 100)
logLikMulti(data, "norm")

}
