% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{GenerateDataBest}
\alias{GenerateDataBest}
\title{Generate multiple samples from a fitted model with specified best performing
model}
\usage{
GenerateDataBest(n, true, parameters, best, models, N, ic = "AIC", ...)
}
\value{
: A list containing the following:.
}
\description{
Generate multiple samples from a fitted model with specified best performing
model
}
\examples{
GenerateData(25, "norm", c(mu = 0.3, sd = 1.2))
}
