% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_frequencies.R
\name{ModelFrequencies}
\alias{ModelFrequencies}
\title{Generate Bootstrap Information Criterion Frequencies}
\usage{
ModelFrequencies(n, true, parameters, models, N = 1000, ic = "AIC")
}
\arguments{
\item{n:}{Sample size.}

\item{true:}{A string giving the name of the model to generate data from.}

\item{parameters:}{A vector of parameterseters specifying a fitted true model.}

\item{models:}{A string vector of model names for selecting between.}

\item{N:}{Number of bootstrap samples.}

\item{ic:}{String giving which information criterion to use (e.g. 'AICc', 'BIC')}
}
\value{
: A list containing the following:.
}
\description{
Generate Bootstrap Information Criterion Frequencies
}
\examples{
models = c("norm0", "norm1", "norm")
ModelFrequencies(25, "norm", c(0.3, 1.2), models, 1000, "AIC")

}
